requirements.txt
# Core dependencies
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.0
einops>=0.6.0
transformers-stream-generator>=0.0.4

# Data processing
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
opencv-python>=4.8.0

# Visualization and monitoring
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
tensorboard>=2.13.0
wandb>=0.15.0

# Web interface
gradio>=4.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Evaluation
torchmetrics>=1.0.0
